# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CIPLA Defeat Hepatitis Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /managers:
    get:
      operationId: listManagers
      tags: [managers]
      summary: List all managers
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: region
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of managers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerListResponse"
    post:
      operationId: createManager
      tags: [managers]
      summary: Create manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
  /managers/{id}:
    get:
      operationId: getManager
      tags: [managers]
      summary: Get manager by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Manager
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
    put:
      operationId: updateManager
      tags: [managers]
      summary: Update manager
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
    delete:
      operationId: deleteManager
      tags: [managers]
      summary: Delete manager
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /managers/{id}/toggle-status:
    post:
      operationId: toggleManagerStatus
      tags: [managers]
      summary: Activate or deactivate a manager
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manager"
  /managers/{id}/progress:
    get:
      operationId: getManagerProgress
      tags: [managers]
      summary: Get manager progress stats
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Manager progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerProgress"
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: managerId
          in: query
          schema: { type: integer }
        - name: doctorId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResponse"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
...
components:
  schemas:
...
    Manager:
      type: object
      required: [id, name, employeeCode, email, mobile, region, headquarters, username, isActive, targetDoctors, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        email:
          type: string
        mobile:
          type: string
        region:
          type: string
        headquarters:
          type: string
        username:
          type: string
        isActive:
          type: boolean
        targetDoctors:
          type: integer
        createdAt:
          type: string
    ManagerInput:
      type: object
      required: [name, employeeCode, email, mobile, region, headquarters, username, password, targetDoctors]
      properties:
        name:
          type: string
        employeeCode:
          type: string
        email:
          type: string
        mobile:
          type: string
        region:
          type: string
        headquarters:
          type: string
        username:
          type: string
        password:
          type: string
        targetDoctors:
          type: integer
    ManagerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        region:
          type: string
        headquarters:
          type: string
        targetDoctors:
          type: integer
        isActive:
          type: boolean
    ManagerListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Manager"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ManagerProgress:
      type: object
      required: [managerId, managerName, target, completed, pending, failed, progressPercent]
      properties:
        managerId:
          type: integer
        managerName:
          type: string
        target:
          type: integer
        completed:
          type: integer
        pending:
          type: integer
        failed:
          type: integer
        progressPercent:
          type: number
...
    Video:
      type: object
      required: [id, doctorId, doctorName, status, createdAt]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        managerId:
          type: integer
        managerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    VideoListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...