# lib/api-zod/src/generated/types/managerListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CIPLA Defeat Hepatitis Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Manager } from './manager';

export interface ManagerListResponse {
  data: Manager[];
  total: number;
  page: number;
  limit: number;
}