# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CIPLA Defeat Hepatitis Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from './videoStatus';

export interface Video {
  id: number;
  doctorId: number;
  doctorName: string;
  managerId?: number;
  /** @nullable */
  managerName?: string | null;
  status: VideoStatus;
  /** @nullable */
  videoUrl?: string | null;
  createdAt: string;
  updatedAt?: string;
}