# lib/api-zod/src/generated/types/videoListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CIPLA Defeat Hepatitis Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Video } from './video';

export interface VideoListResponse {
  data: Video[];
  total: number;
  page: number;
  limit: number;
}